Using new collisions
Collisions are based on intersections between wall rects and car rects.
The collisions on corners seem more realistic and there are less known
points to traverse walls. The downside is that on the right hand side of
walls there are issues but these seem less of a problem than the old
issues.